# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loreline",
    "summary": "Tabletop session transcriber - capture + STT orchestration.",
    "version": "1"
  },
  "paths": {
...
    "/api/system/diarizer/probe": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Probe Diarizer Endpoint",
        "description": "Probe an arbitrary diarization endpoint, uncached and on demand.\n\nDeliberately not routed through ``_diarizer_status`` above: that cache\nexists because ``/healthz`` is polled every few seconds for one settings\nvalue, while this answers a one-off question about whatever a GM is\ncurrently typing into the capture panel's endpoint field - a different\nstring on every keystroke, which a cache keyed for a single settings value\nwould not help and could even answer wrong (a stale verdict for a value\nthat has since changed back).\n\nRequires auth like every other route here: naming an endpoint makes the\nserver issue an arbitrary outbound HTTP request, which is already true of\nthe stored default's probe behind ``/healthz`` above - this is the same\nexposure on a second, caller-supplied value, not a new one.",
        "operationId": "probe_diarizer_endpoint_api_system_diarizer_probe_get",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "endpoint",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Endpoint"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiarizerProbeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/revision": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Revision",
        "description": "Return the deployed commit, `git describe`'s name for it, and the way back.\n\nBoth, rather than one: the SHA is what /rollback takes and what the update\nresult reports, while the described string is the only one of the two worth\nreading off a screen. Either can be null where this deployment cannot know\nit - a Docker image built without the revision baked in - and the UI is\nwritten to show a dash for that instead of a wrong answer.\n\nThe previous pair is the same two answers for the commit HEAD was on before\nit last moved, which is what Settings > Client offers to roll back to, and\n``rollback_unavailable`` is the one sentence a Docker deployment shows in\nplace of that offer. Answered here rather than on a route of its own so the\npage learns everything it needs about the deployment in one call.",
        "operationId": "revision_api_system_revision_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/system/update": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Trigger Update",
        "description": "Run the self-update: git pull + uv sync + restart, or the updater service in Docker.",
        "operationId": "trigger_update_api_system_update_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/system/rollback": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Rollback",
        "description": "Roll the deployment back to a prior commit.",
        "operationId": "rollback_api_system_rollback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/system/autostart": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get Autostart",
        "description": "Report whether the systemd unit is enabled at boot.",
        "operationId": "get_autostart_api_system_autostart_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutostartState"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "put": {
        "tags": [
          "system"
        ],
        "summary": "Set Autostart",
        "description": "Enable or disable systemd autostart.",
        "operationId": "set_autostart_api_system_autostart_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AutostartUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AutostartState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/system/defaults": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "Get Defaults",
        "description": "Return the per-action default models/mode used to pre-select the pickers.\n\nA blank stored summary or recap prompt is served as the built-in default\ntext, so the settings UI always shows the concrete, editable instructions -\nclearing the field and saving is the reset-to-default gesture.",
        "operationId": "get_defaults_api_system_defaults_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionDefaults"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "put": {
        "tags": [
          "system"
        ],
        "summary": "Set Defaults",
        "description": "Persist the per-action defaults.\n\nA summary or recap prompt equal to the built-in default (or blank) is\nstored blank, so an untouched field keeps tracking future improvements to\nthe built-in text instead of pinning today's copy. The response mirrors\nGET: served filled in.",
        "operationId": "set_defaults_api_system_defaults_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionDefaults"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionDefaults"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/system/alerts/channels": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "List Alert Channels",
        "description": "List configured alert channels (tokens masked as set/unset).",
        "operationId": "list_alert_channels_api_system_alerts_channels_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertChannelView"
                  },
                  "type": "array",
                  "title": "Response List Alert Channels Api System Alerts Channels Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Create Alert Channel",
        "description": "Add an alert channel; store its token write-only when supplied.",
        "operationId": "create_alert_channel_api_system_alerts_channels_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertChannelWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertChannelView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/system/alerts/channels/{channel_id}/test": {
      "post": {
        "tags": [
          "system"
        ],
        "summary": "Test Alert Channel",
        "description": "Send a test notification to one channel, and say why when it fails.\n\nAlways 200, failure included: the request itself succeeded, and the answer\nto \"did this channel take it\" is the body. Turning a refused webhook into a\n5xx here would make the page's own error path swallow the reason.",
        "operationId": "test_alert_channel_api_system_alerts_channels__channel_id__test_post",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "channel_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Channel Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertTestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/system/services": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "List Services",
        "description": "The containers in this compose project, with their current state.",
        "operationId": "list_services_api_system_services_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServiceState"
                  },
                  "type": "array",
                  "title": "Response List Services Api System Services Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiarizerProbeResponse": {
        "properties": {
          "reachable": {
            "type": "boolean",
            "title": "Reachable"
          },
          "status": {
            "$ref": "#/components/schemas/HealthStatus"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "reachable",
          "status"
        ],
        "title": "DiarizerProbeResponse",
        "description": "The graded probe of one diarization endpoint, checked on demand.\n\nThe same verdict shape ``HealthResponse`` already carries for the stored\ndefault, but for whatever endpoint the caller names. The capture panel\nuses this to check the value currently typed into its endpoint field,\nwhich ``/healthz`` cannot: that call only ever probes\n``defaults.diar_endpoint``, the last *saved* value, never a value someone\nis only trying out."
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "description": "Session JWT set by `POST /api/auth/login`. The cookie is HttpOnly, so pasting a value into this dialog does nothing: browsers drop the `Cookie` header Swagger UI would set from it. To use `Try it out`, run `POST /api/auth/login` from this page first (or log into the app in the same browser); the cookie is then sent automatically. Outside a browser, send the cookie yourself, e.g. `curl -c jar -X POST .../api/auth/login -d '{\"password\":\"...\"}' -H 'content-type: application/json'` then `curl -b jar .../api/session`. When no password is configured auth is off entirely and the cookie is not required.",
        "in": "cookie",
        "name": "loreline_token"
      }
    }
  }
}